backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
arize-otel>=0.8.1
aiohttp>=3.9.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
openinference-instrumentation-langchain>=0.1.19
openinference-instrumentation-openai>=0.1.0
openinference-instrumentation-litellm>=0.1.0
openinference-instrumentation>=0.1.12
langgraph>=0.2.55
langchain>=0.3.7
langchain-openai>=0.2.10
langchain-community>=0.3.5
litellm
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
